Mostly complete expression parsing
This is still lacking proper error handling and I have left some TODOs on the
code to deal with other improvements. The main one I am concerned with here is
using the BitSet.

The reason for using a bitset came from thinking about implementing
`next-execution-time` for a cron. I guess that I would need to skip around the
selected values for each cron segment and a bitset provides a nice and efficient
interface for that.