Use BitSet to mark trigger slots
This removes any use of hash-set and `*` is now handled in a more uniform way. I
have also unified the numeric and named parsing into one. we might be doing some
extra work for segments that only had numeric values but it's minimal and well
handled so I would rather keep things unified.