Abstract expression parsing
Hid it behide a simple to use macro and started making use of transducer. My
concern here is that creating the transducers on every call is a bit inefficient
though that should not matter much, at least I will ignore it until it becomes a
n issue. Parsing `* * * * * *` right now is taking a tenth of a millisecond.