Implement `next-execution`
This is a heuristic I came up with and eventually converges to a correct value,
at least I hope so, tests will be needed here for you, along with a way to
provide a starting timestamp.

The idea is that we keep moving towards the right direction, it's quite
straightforward except for handling weekdays. There, if everything else has been
shifted properly then we jiggle things to the right day and left things converge again.